9742: moving computation of trial info to computeds
