+++ web-client/src/presenter/computeds/formattedMessages.js
@@ -15,6 +15,20 @@ export const formattedMessages = (get, applicationContext) => {
     tableSort,
   });
 
+  const { STATUS_TYPES } = applicationContext.getConstants();
+
+  messages.forEach(message => {
+    message.showTrialInformation =
+      message.caseStatus === STATUS_TYPES.calendared;
+
+    if (message.showTrialInformation) {
+      setTrialInformationOnMessage({
+        applicationContext,
+        message,
+      });
+    }
+  });
+
   const { box } = get(state.messageBoxToDisplay);
   const { role } = get(state.user);
   const { USER_ROLES } = applicationContext.getConstants();
@@ -55,4 +69,23 @@ export const formattedMessages = (get, applicationContext) => {
   }
 
   return sharedComputedResult;
+};
+
+export const setTrialInformationOnMessage = ({
+  applicationContext,
+  message,
+}) => {
+  const { TRIAL_SESSION_SCOPE_TYPES } = applicationContext.getConstants();
+
+  if (message.trialLocation !== TRIAL_SESSION_SCOPE_TYPES.standaloneRemote) {
+    message.formattedTrialLocation = applicationContext
+      .getUtilities()
+      .abbreviateState(message.trialLocation);
+  } else {
+    message.formattedTrialLocation = message.trialLocation;
+  }
+
+  message.formattedTrialDate = applicationContext
+    .getUtilities()
+    .formatDateString(message.trialDate, 'MMDDYY');
 };
