+++ web-client/src/presenter/computeds/formattedMessages.test.js
@@ -1,9 +1,12 @@
 /* eslint-disable max-lines */
+import { applicationContextForClient as applicationContext } from '../../../../shared/src/business/test/createTestApplicationContext';
 import { formattedMessages as formattedMessagesComputed } from './formattedMessages';
 import { runCompute } from 'cerebral/test';
 import { withAppContextDecorator } from '../../withAppContext';
 
 describe('formattedMessages', () => {
+  const { STATUS_TYPES, TRIAL_SESSION_SCOPE_TYPES } =
+    applicationContext.getConstants();
   const formattedMessages = withAppContextDecorator(formattedMessagesComputed);
 
   it('returns filtered messages sorted oldest to newest and completedMessages from state.messages when messageBoxToDisplay.box is inbox', () => {
@@ -487,5 +490,118 @@ describe('formattedMessages', () => {
     });
 
     expect(result.messages.length).toEqual(2);
+  });
+
+  it('should set showTrialInformation to true when caseStatus is calendared', () => {
+    const mockCalendaredMessage = {
+      caseStatus: STATUS_TYPES.calendared,
+      completedAt: '2019-01-02T16:29:13.122Z',
+      createdAt: '2019-01-01T16:29:13.122Z',
+      docketNumber: '101-20',
+      message: 'This is a test message',
+      trialDate: '2025-01-01T16:29:13.122Z',
+      trialLocation: 'Austin, TX',
+    };
+    const result = runCompute(formattedMessages, {
+      state: {
+        messageBoxToDisplay: {
+          box: 'outbox',
+        },
+        messages: [mockCalendaredMessage],
+        screenMetadata: {},
+        user: {
+          role: 'adc',
+        },
+      },
+    });
+
+    expect(result.messages[0].showTrialInformation).toBe(true);
+  });
+
+  it('should set showTrialInformation to false when caseStatus is NOT calendared', () => {
+    const mockCalendaredMessage = {
+      caseStatus: STATUS_TYPES.new,
+      completedAt: '2019-01-02T16:29:13.122Z',
+      createdAt: '2019-01-01T16:29:13.122Z',
+      docketNumber: '101-20',
+      message: 'This is a test message',
+    };
+
+    const result = runCompute(formattedMessages, {
+      state: {
+        messageBoxToDisplay: {
+          box: 'outbox',
+        },
+        messages: [mockCalendaredMessage],
+        screenMetadata: {},
+        user: {
+          role: 'adc',
+        },
+      },
+    });
+
+    expect(result.messages[0].showTrialInformation).toBe(false);
+  });
+
+  it('should format the trialDate and trialLocation on the message when caseStatus is Calendared', () => {
+    const mockCalendaredMessage = {
+      caseStatus: STATUS_TYPES.calendared,
+      completedAt: '2019-01-02T16:29:13.122Z',
+      createdAt: '2019-01-01T16:29:13.122Z',
+      docketNumber: '101-20',
+      message: 'This is a test message',
+      trialDate: '2019-01-01T16:29:13.122Z',
+      trialLocation: 'Houston, Texas',
+    };
+
+    const result = runCompute(formattedMessages, {
+      state: {
+        messageBoxToDisplay: {
+          box: 'outbox',
+        },
+        messages: [mockCalendaredMessage],
+        screenMetadata: {},
+        user: {
+          role: 'adc',
+        },
+      },
+    });
+
+    expect(result.messages[0]).toMatchObject({
+      formattedTrialDate: '01/01/19',
+      formattedTrialLocation: 'Houston, TX',
+    });
+  });
+
+  it(`should not abbreviate trialLocation when it is ${TRIAL_SESSION_SCOPE_TYPES.standaloneRemote}`, () => {
+    const mockCalendaredMessage = {
+      caseStatus: STATUS_TYPES.calendared,
+      completedAt: '2019-01-02T16:29:13.122Z',
+      createdAt: '2019-01-01T16:29:13.122Z',
+      docketNumber: '101-20',
+      message: 'This is a test message',
+      trialDate: '2019-01-01T16:29:13.122Z',
+      trialLocation: TRIAL_SESSION_SCOPE_TYPES.standaloneRemote,
+    };
+
+    const result = runCompute(formattedMessages, {
+      state: {
+        messageBoxToDisplay: {
+          box: 'outbox',
+        },
+        messages: [mockCalendaredMessage],
+        screenMetadata: {},
+        user: {
+          role: 'adc',
+        },
+      },
+    });
+
+    expect(
+      applicationContext.getUtilities().abbreviateState,
+    ).not.toHaveBeenCalled();
+    expect(result.messages[0]).toMatchObject({
+      formattedTrialLocation: TRIAL_SESSION_SCOPE_TYPES.standaloneRemote,
+    });
   });
 });
