+++ web-client/src/presenter/utilities/processFormattedMessages.js
@@ -73,19 +73,6 @@ export const getFormattedMessages = ({
       }),
       'MMDDYY',
     );
-    const { STATUS_TYPES } = applicationContext.getConstants();
-
-    messages.forEach(message => {
-      message.showTrialInformation =
-        message.caseStatus === STATUS_TYPES.calendared;
-
-      if (message.showTrialInformation) {
-        setTrialInformationOnMessage({
-          applicationContext,
-          message,
-        });
-      }
-    });
 
     const formattedMessages = messages.map(message => {
       const {
@@ -196,20 +183,4 @@ export const applyFiltersToCompletedMessages = ({
     },
     filteredCompletedMessages,
   };
-};
-
-const setTrialInformationOnMessage = ({ applicationContext, message }) => {
-  const { TRIAL_SESSION_SCOPE_TYPES } = applicationContext.getConstants();
-
-  if (message.trialLocation !== TRIAL_SESSION_SCOPE_TYPES.standaloneRemote) {
-    message.formattedTrialLocation = applicationContext
-      .getUtilities()
-      .abbreviateState(message.trialLocation);
-  } else {
-    message.formattedTrialLocation = message.trialLocation;
-  }
-
-  message.formattedTrialDate = applicationContext
-    .getUtilities()
-    .formatDateString(message.trialDate, 'MMDDYY');
 };
